+++ extension/entity.h
@@ -5,8 +5,8 @@
 
 typedef struct Entity
 {
-    Vec2f pos;
-    Vec2f velocity;
+    Vec2d pos;
+    Vec2d velocity;
 
     // relative to pos
     Rectangle hitbox;
@@ -15,17 +15,17 @@ typedef struct Entity
 
     int HP, maxHP;
     float ATK;
-    float SPD;
+    double SPD;
 
 } Entity;
 
 Entity Entity_construct();
 
-Entity Entity_construct_generic(Rectangle, Vec2f);
+Entity Entity_construct_generic(Rectangle, Vec2d);
 
 typedef struct Player
 {
     Entity entity;
-    float acceleration_const; // between 0 and 1
+    double acceleration_const; // between 0 and 1
 } Player;
 #endif // ENTITY_H
